requirements.txt
selenium
psutil
pyyaml
